final beta v1

+++ public/src/ctrl/signInCtrl.js
@@ -4,11 +4,11 @@
     angular
         .module('funfun')
         .controller('signInCtrl', [
-            '$scope', '$mdDialog', '$firebaseAuth', '$location', 'localStorageService', '$route',
+            '$scope', '$mdDialog', '$firebaseAuth', '$location', 'localStorageService', '$route','SweetAlert',
             signInCtrl
         ])
 
-    function signInCtrl($scope, $mdDialog, $firebaseAuth, $location, localStorageService, $route) {
+    function signInCtrl($scope, $mdDialog, $firebaseAuth, $location, localStorageService, $route, SweetAlert) {
 
         $scope.signIn = function(data, text) {
             var t1 = 'email'
@@ -21,19 +21,21 @@
                 if (email == "admin@system" && password == "p@ssW0rd") {
                     console.log("Hello Admin")
                     localStorageService.set("checkAdmin", text)
-                    $('#signin').modal('hide')
+                    $mdDialog.hide()
+                    location.reload();
                     $location.path('/admin');
                 } else {
                     auth.$signInWithEmailAndPassword(email, password).then(function(firebaseUser) {
                         checkLoginFromDb(firebaseUser)
-                            // Add Sweet  Alert
+                        // Add Sweet  Alert
                     }).catch(function(error) {
                         $mdDialog.hide()
                         var errorCode = error.code
                         var errorMessage = error.message
-                            // [START_EXCLUDE]
+                        // [START_EXCLUDE]
+                          SweetAlert.swal("กรุณาสมัครสมาชิก หรือ email และ password ไม่ถูกต้อง!", error.message , "error")
                         console.log("Please SignUp to Website Authentication failed:", error)
-                            // ...
+                        // ...
                     })
                 }
             } else {
@@ -47,23 +49,29 @@
             }
         }
         $scope.logout = function() {
-            firebase.auth().signOut().then(function() {
-                localStorageService.remove("status")
+            if (localStorageService.get("checkAdmin") != undefined) {
                 localStorageService.remove("checkAdmin")
                 $location.path("/")
-            }).catch(function(error) {
-                console.log("Authentication failed:", error)
-            })
+                location.reload();
+            } else {
+                firebase.auth().signOut().then(function() {
+                    localStorageService.remove("status")
+                    $location.path("/")
+                    location.reload();
+                }).catch(function(error) {
+                    console.log("Authentication failed:", error)
+                })
+            }
         }
 
         function checkLoginFromDb(firebaseUser) {
             if (firebaseUser.uid == null) {
                 var uid = firebaseUser.user.uid
-                    //console.log(uid) //debug uid from firebase
+                //console.log(uid) //debug uid from firebase
                 getUserForCheckLogin(uid)
             } else {
                 var uid = firebaseUser.uid
-                    //console.log(uid) //debug uid from firebase
+                //console.log(uid) //debug uid from firebase
                 getUserForCheckLogin(uid)
             }
         }
@@ -71,11 +79,13 @@
         function getUserForCheckLogin(paramId) {
             firebase.database().ref('users/' + paramId).once('value').then(function(snapshot) {
                 var status = snapshot.val().status
-                    // Create localStorage for Session
+                // Create localStorage for Session
                 localStorageService.set("status", status)
                 console.log(status)
                 $mdDialog.hide()
+                location.reload()
             }).catch(function(error) {
+                SweetAlert.swal("ชื่อบัญชีนี้ไม่ถูกต้อง! กรุณาสมัครสมาชิก", error.message , "error")
                 console.log("Please SignUp to Website Authentication failed:", error)
                 firebase.auth().signOut().then(function() {
                     localStorageService.clearAll()
